chore: move CertWatcher to HttpsServer.hpp
